spec-366: extract tool-specs.ts handler infra into handlers/shared.ts
sol-1 of audit spec-345 (umbrella spec-354), step 1: lift the shared
handler infrastructure (ToolCtx/ToolSpec/ResolvedRef/FooterSignal types,
the reqCtx bridge, and the rendering + guidance-envelope helpers) out of
agent/tool-specs.ts into agent/handlers/shared.ts. tool-specs.ts imports
the infra from there and re-exports every symbol external modules/tests
import from it, so no import site moved.

Behaviour-preserving: tool contract (std-16) unchanged; tsc clean (only
the 14 pre-existing unrelated test errors remain); tool-specs integration
+ audit + b-97 suites green; legacy-spec-vocab guard green.
